# server/docs/openapi.yaml
openapi: 3.0.0
info:
  contact: {}
  description: Feature flag system with rule-based evaluation, segment targeting,
    pack-based feature bundling, and A/B experiments.
  title: Feature Evaluator API
  version: "1.0"
paths:
...
  /admin/experiments:
    get:
      description: Returns all experiments in the current workspace
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Da\
                  taResponse-array_github_com_rendis_feature-evaluator_internal\
                  _dto_ExperimentResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
      security:
        - BearerAuth: []
      summary: List experiments
      tags:
        - experiments
    post:
      description: Creates a new A/B experiment for a feature
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Cr\
                eateExperimentRequest"
        description: Experiment creation payload
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Ex\
                  perimentResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
      security:
        - BearerAuth: []
      summary: Create experiment
      tags:
        - experiments
  "/admin/experiments/{id}":
    get:
      description: Returns a single experiment by ID
      parameters:
        - description: Experiment ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Ex\
                  perimentResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
      security:
        - BearerAuth: []
      summary: Get experiment
      tags:
        - experiments
    put:
      description: Updates a draft experiment by ID
      parameters:
        - description: Experiment ID
          in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Up\
                dateExperimentRequest"
        description: Experiment update payload
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Ex\
                  perimentResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
      security:
        - BearerAuth: []
      summary: Update experiment
      tags:
        - experiments
  "/admin/experiments/{id}/complete":
    post:
      description: Completes a running or paused experiment
      parameters:
        - description: Experiment ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Me\
                  ssageResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
      security:
        - BearerAuth: []
      summary: Complete experiment
      tags:
        - experiments
  "/admin/experiments/{id}/declare-winner":
    post:
      description: Declares a winning variant for a completed experiment
      parameters:
        - description: Experiment ID
          in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.De\
                clareWinnerRequest"
        description: Winner declaration payload
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.De\
                  clareWinnerResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
      security:
        - BearerAuth: []
      summary: Declare winner
      tags:
        - experiments
  "/admin/experiments/{id}/pause":
    post:
      description: Pauses a running experiment
      parameters:
        - description: Experiment ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Me\
                  ssageResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
      security:
        - BearerAuth: []
      summary: Pause experiment
      tags:
        - experiments
  "/admin/experiments/{id}/results":
    get:
      description: Returns computed results and statistics for an experiment
      parameters:
        - description: Experiment ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_domain\
                  _experiment.Results"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
      security:
        - BearerAuth: []
      summary: Get experiment results
      tags:
        - experiments
  "/admin/experiments/{id}/start":
    post:
      description: Starts a draft or paused experiment
      parameters:
        - description: Experiment ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Me\
                  ssageResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
      security:
        - BearerAuth: []
      summary: Start experiment
      tags:
        - experiments
...
  /eval/conversions:
    post:
      description: Records a conversion event for an experiment metric
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Re\
                cordConversionRequest"
        description: Conversion event payload
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Me\
                  ssageResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      summary: Record conversion
      tags:
        - experiments
...
servers:
  - url: /features
components:
  securitySchemes:
    ApiKeyAuth:
      description: API key for evaluation or admin access
      in: header
      name: X-Api-Key
      type: apiKey
    BearerAuth:
      description: Type "Bearer" followed by a space and JWT token
      in: header
      name: Authorization
      type: apiKey
    WorkspaceHeader:
      description: Workspace key for multi-tenant isolation
      in: header
      name: X-Workspace
      type: apiKey
...